战斗基础扣血循环

+++ Assets/Scripts/Runtime/Config/Cards/NormalCardConfig.cs
@@ -14,6 +14,6 @@ namespace Config
     {
         public string cardId;
         public string cardImgPath;
-        public float cardPoint;
+        public int cardPoint;
     }
 }
